refactor(states): break circular imports between states and engines
Extract IEntityState, Heading, and getPose into dedicated modules and
switch entity/player/physicsUtils imports to direct paths so states no
longer barrel-import engines. Public exports through states/index.ts are
unchanged aside from the structural split.
